Simplify event store logging
